+++ packages/react/index.js
@@ -1,0 +1,4 @@
+export const adapter = {
+  framework: "react",
+  packages: ["eslint-plugin-jsx-a11y"]
+};